fixing URL of single page manual
when clicking Info > Manual in the MainForm but no manual was installed on the system, Hydrogen will redirect to http://hydrogen-music.org/documentation/manual/manual_en.html as a fallback.

When introducing the single page and the chunked versions the link above was not used the browser visit a dead link. I made the single page version the default one and used manual_en.html for it.